requirements.txt
colorama
json5
Pillow